src/construction/over-array.ts
/**
 * @packageDocumentation
 * @module @proc7ts/push-iterator
 */
import { arrayIterator } from '../array-iterator.impl';
import type { PushIterable } from '../push-iterable';
import type { PushIterator } from '../push-iterator';
import { PushIterator__symbol } from '../push-iterator';

/**
 * Creates a {@link PushIterable push iterable} over elements of array-like structure.
 *
 * @typeParam T  Array elements type.
 * @param array  An array-like structure. E.g. `Array`, DOM `NodeList`, etc.
 *
 * @returns New push iterable over array elements.
 */
export function overArray<T>(array: ArrayLike<T>): PushIterable<T> {

  const iterate = (): PushIterator<T> => arrayIterator(array);

  return {
    [Symbol.iterator]: iterate,
    [PushIterator__symbol]: iterate,
  };
}
